Moved classes to correct packages
Moved client classes to correct package.
