+++ 1.9/src/main/java/code/elix_x/coremods/colorfulblocks/client/gui/GuiSelectColor.java
@@ -1,4 +1,4 @@
-package code.elix_x.coremods.colorfulblocks.gui;
+package code.elix_x.coremods.colorfulblocks.client.gui;
 
 import code.elix_x.coremods.colorfulblocks.ColourfulBlocksBase;
 import code.elix_x.coremods.colorfulblocks.net.ColorChangeMessage;
