+++ 1.9/src/main/java/code/elix_x/coremods/colorfulblocks/proxy/ClientProxy.java
@@ -18,11 +18,11 @@ import code.elix_x.coremods.colorfulblocks.ColourfulBlocksBase;
 import code.elix_x.coremods.colorfulblocks.client.color.ColorfulBlocksBlockColor;
 import code.elix_x.coremods.colorfulblocks.client.color.StandartColoringToolItemColor;
 import code.elix_x.coremods.colorfulblocks.client.events.LastRenderWorldEvent;
+import code.elix_x.coremods.colorfulblocks.client.gui.GuiSelectColor;
 import code.elix_x.coremods.colorfulblocks.color.material.ColoringMaterialsManager;
 import code.elix_x.coremods.colorfulblocks.color.tool.ColoringToolProvider;
 import code.elix_x.coremods.colorfulblocks.color.tool.ColoringToolsManager;
 import code.elix_x.coremods.colorfulblocks.color.tool.IColoringTool;
-import code.elix_x.coremods.colorfulblocks.gui.GuiSelectColor;
 import code.elix_x.excore.utils.reflection.AdvancedReflectionHelper.AField;
 import net.minecraft.block.Block;
 import net.minecraft.client.Minecraft;
